core: Move pager.rs to storage module
