+++ core/lib.rs
@@ -3,10 +3,8 @@ mod datetime;
 mod error;
 mod function;
 mod io;
-mod pager;
 mod pseudo;
 mod schema;
-mod sqlite3_ondisk;
 mod storage;
 mod translate;
 mod types;
@@ -20,12 +18,12 @@ static GLOBAL: mimalloc::MiMalloc = mimalloc::MiMalloc;
 
 use fallible_iterator::FallibleIterator;
 use log::trace;
-use pager::Pager;
 use schema::Schema;
-use sqlite3_ondisk::DatabaseHeader;
 use sqlite3_parser::{ast::Cmd, lexer::sql::Parser};
 use std::sync::Arc;
 use std::{cell::RefCell, rc::Rc};
+use storage::pager::Pager;
+use storage::sqlite3_ondisk::DatabaseHeader;
 
 pub use error::LimboError;
 pub type Result<T> = std::result::Result<T, error::LimboError>;
