+++ core/storage/pager.rs
@@ -1,6 +1,5 @@
 use crate::buffer_pool::BufferPool;
-use crate::sqlite3_ondisk::PageContent;
-use crate::sqlite3_ondisk::{self, DatabaseHeader};
+use crate::storage::sqlite3_ondisk::{self, DatabaseHeader, PageContent};
 use crate::wal::Wal;
 use crate::{Buffer, PageSource, Result};
 use log::trace;
